Rename render folder to rendering
Prepare to move anything related to rendering into this direction
