+++ src/rendering/templates/bookTemplate.njk
