+++ src/rendering/templates/defaultHighlightTemplate.njk
