+++ src/rendering/templates/highlightTemplateWrapper.njk
